styling and assignment page buttons
restyled page to make prompt cards larger and to make input fileds one line. This will help keep the size of the prompt box within the window. Also created a mapping of projects to assignment pages and changed the timeline item component to only add assignment apge buttons if there is a link